insert coin

+++ VendingMachineCore/Products/Product.cs
@@ -2,7 +2,7 @@ namespace VendingMachineCore.Products
 {
     public abstract class Product
     {
-        protected double Prize { get; set; }
+        protected int Prize { get; set; }
         public double Number { get; set; }
 
         public Product()
